+++ .gitignore
@@ -33,4 +33,5 @@
 
 # build
 build/
+build2/
 Testing/
