+++ test_expressions/CMakeLists.txt
@@ -10,3 +10,13 @@ target_link_libraries(test_expressions
 )
 gtest_discover_tests(test_expressions)
 set_target_properties(test_expressions PROPERTIES CXX_CLANG_TIDY "")
+
+add_executable(benchmark_expressions
+    benchmark_expressions.cc
+)
+target_link_libraries(benchmark_expressions
+    glog::glog
+    ky_expressions
+    expressions2
+    benchmark::benchmark
+)
